# backend/empresa/target/classes/api.yaml
openapi: "3.0.3"
info:
  title: Biblioteca API
  description: API para gerenciar organizações
  version: 1.0.0
...
paths:
...
  /usuario/find:
    post:
      summary: Busca usuários
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: findUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRequest'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponseDTO'
  /usuario/detail/{id}:
    get:
      summary: Detalha  um usuário
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: detailUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponseDTO'
  /usuario/create:
    put:
      summary: Cria um usuário
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: createUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioModel'
  /usuario/update:
    patch:
      summary: Atualiza um usuário
      security:
          - bearerAuth: []
      tags:
        - Usuario
      operationId: updateUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioModel'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioModel'
  /usuario/delete/{id}:
    delete:
      summary: Deleta um usuário
      security:
          - bearerAuth: []
      tags:
        - Usuario
      operationId: deleteUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
  /usuario/foto/{usuario_id}/{ativo}:
    get:
      summary: Busca a foto de perfil do usuário
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: getFotoUsuario
      parameters:
        - name: usuario_id
          in: path
          required: true
          schema:
            type: integer
          description:  Id do usuário
        - name: ativo
          in: path
          required: true
          schema:
              type: boolean
          description: Indica se a foto está ativa
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioFotoModel'
  /usuario/foto:
    put:
      summary: Insere foto do usuário
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: createFotoUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioFotoModel'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioFotoModel'
    patch:
      summary: Atualiza foto do usuário
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: updateFotoUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioFotoModel'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioFotoModel'
    delete:
      summary: Deleta foto do usuário
      security:
        - bearerAuth: []
      tags:
        - Usuario
      operationId: deleteFotoUsuario
      parameters:
        - name: usuario_id
          in: path
          required: true
          schema:
            type: integer
          description:  Id do usuário
      responses:
        '204':
          description: No Content
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # Especifica o formato do token, pode ser omitido
  schemas:
...
    UsuarioRequest:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        dataNascimento:
          type: string
          format: date
        perfis:
          type: array
          items:
            type: integer
        limit:
            type: integer
        offset:
          type: integer
        event:
          $ref: '#/components/schemas/EventPagination'
    EventPagination:
      type: object
      properties:
        length:
          type: integer
        pageIndex:
          type: integer
        pageSize:
            type: integer
        previousPageIndex:
            type: integer
...
    UsuarioResponseDTO:
      type: object
      properties:
        usuarios:
          type: array
          items:
            $ref: '#/components/schemas/UsuarioModel'
        totalRecords:
          type: integer
...
    RegisterRequestDTO:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        senha:
          type: string
        dataNascimento:
          type: string
          format: date
        enderecos:
          type: array
          items:
            $ref: '#/components/schemas/enderecoModel'
        perfis:
          type: array
          items:
            $ref: '#/components/schemas/PerfilModel'
        foto_perfil:
          type: string
          format: byte
        foto:
          $ref: '#/components/schemas/UsuarioFotoModel'
...
    UsuarioModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        senha:
          type: string
        email:
          type: string
        dataNascimento:
          type: string
          format: date
        enderecos:
          type: array
          items:
            $ref: '#/components/schemas/enderecoModel'
        perfis:
            type: array
            items:
                $ref: '#/components/schemas/PerfilModel'
        foto:
          $ref: '#/components/schemas/UsuarioFotoModel'
    UsuarioFotoModel:
      type: object
      properties:
        id:
          type: integer
        usuario_id:
          type: integer
        foto:
          type: string
          format: byte
        ativo:
          type: boolean
        data_criacao:
          type: string
          format: date-time
        data_atualizacao:
          type: string
          format: date-time
    enderecoModel:
      type: object
      properties:
        id:
          type: integer
        cidade_id:
          $ref: '#/components/schemas/CidadeModel'
        cep:
          type: string
        logradouro:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        latitude:
          type: number
        longitude:
            type: number
        usuario_id:
          type: integer
    PerfilModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
    CidadeModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        estado_id:
          $ref: '#/components/schemas/EstadoModel'
    EstadoModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        pais_id:
          $ref: '#/components/schemas/PaisModel'
    PaisModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
...